requirements.txt
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
cycler==0.10.0
decorator==4.4.2
future==0.18.2
idna==2.9
kiwisolver==1.1.0
matplotlib==3.2.1
numpy==1.18.2
pycparser==2.20
pyparsing==2.4.6
python-dateutil==2.8.1
python-telegram-bot==12.4.2
requests==2.23.0
six==1.14.0
tornado==6.0.4
urllib3==1.25.8
